requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
# cryptography added to accomodate openssl version, seems to break snowpark pandas FIGURE THIS OUT

azure-functions
azure-identity
azure-storage-blob
azure-keyvault-secrets
numpy
openpyxl==3.0.9
Pillow==9.2.0
pandas==1.4.1
snowflake-snowpark-python[pandas]==1.0.0
#cryptography==38.0.1